requirements.txt
Django==2.0.7
djangorestframework==3.8.2
django-filter==2.0.0
markdown==2.6.11
psycopg2==2.7.5
django-cors-headers==2.4.0
django-rest-auth[with_social]